+++ src/hooks/useRedux.ts
@@ -1,6 +1,6 @@
 import { useDispatch, useSelector } from 'react-redux';
 
-import type { AppDispatch, RootState } from './store';
+import type { AppDispatch, RootState } from '../redux/store';
 
 const useAppDispatch = useDispatch.withTypes<AppDispatch>();
 const useAppSelector = useSelector.withTypes<RootState>();
