+++ src/main.tsx
@@ -3,8 +3,8 @@ import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { PersistGate } from 'redux-persist/integration/react';
 
-import App from './app/App.tsx';
-import { persistor, store } from './app/store.ts';
+import App from './App.tsx';
+import { persistor, store } from './redux/store.ts';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
