+++ src/redux/rootReducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from '@reduxjs/toolkit';
 
-import authReducer from '../features/auth/authSlice';
+import authReducer from './slices/authSlice';
 
 const rootReducer = combineReducers({
   auth: authReducer,
