+++ src/ts/interfaces.ts
